Split DISPLAY_{TO,CC,BCC} on semicolon to extract multiple recipients
When parsing a "sent" message (instead of "received" one, which contains
actual headers for "To", "Cc" etc.), the "DISPLAY_TO", "DISPLAY_CC" and
"DISPLAY_BCC" properties are used to determine recipients.

The way the data is extracted was incorrect, because according to
Microsoft's documentation[1], the DISPLAY_* properties contain semi-colon-
separated list of display-names of recipients in a single string, while
"email.headerregistry.AddressHeader" expects a list or comma-separated
string of "[display_name] <username@domain>".

[1] https://learn.microsoft.com/en-us/openspecs/exchange_server_protocols/ms-oxprops/777fafff-520d-4141-85e3-9ec41013551d